requirements.txt
openai==1.57.0
Pillow==11.0.0
pytesseract==0.3.13
streamlit==1.40.2
torch==2.5.1
transformers==4.47.0
